# 8349713: [leyden] Memory map the cached code file