@franz1981 2025-02-10
@shipilev 

Does the numbers still holds with
```
sync; echo 3 > /proc/sys/vm/drop_caches
```
before the benchmark (single shot) run?
With big servers, unless there's some expected sharing (i.e. multiple processes using the same archive) to boost it, is that direct I/O could still benefit loading the archive by saving the extra copy (in the OS page cache) required to use the data read.
The other concern re mmap is due to munmap cost which, on kernel side, relies (IIRC) to some v page single (!) lock to guard it - which usually slowdown processes termination