@shipilev 2025-02-10
> Good. Until, as we discussed, cached code (and its data) will be part of CDS archive file.

Yes, that would make this whole thing no-op. CDS mmap-s the archive already.

> How CDS handles `mmap` on windows?

Windows can still memory-map the file through its Windows APIs. But I _vaguely_ recollect some corner cases that @iklam fights every so often. Something with remapping or so? See `MetaspaceShared::use_windows_memory_mapping`, for example. I don't think it is wise to spend time dealing with those corner cases in current Leyden prototype, let it be part of whole CDS archive first.