@shipilev 2025-02-10
> With the Windows APIs used by HotSpot today (VirtualAlloc and MapViewOfFile), we can't map a file into a reserved region. If we just want to mmap the SCCache into non-reserved, random location that's picked by the OS, we can already do that today with HotSpot.
> 
> In some cases, CDS wants to mmap into reserved regions. On Windows, we end up not mapping with MapViewOfFile, but simply reading the entire CDS file into reserved memory.

OK, great. AFAICS, this technically allows us to do `MmapCachedCode = true` on Windows too. Non-Windows currently rely on `MAP_PRIVATE` to get us COW. I believe `FILE_MAP_COPY` in `os::pd_map_memory` gives us the same on Windows.

Still, I think it is a bit saner to keep doing mmap-ing only on Linux, if only to test that non-mmap path works and not force us to debug Windows mmap issues in Leyden prototype. But I don't feel strongly about this. Opinions welcome!