@shipilev 2025-02-11
Any other opinions about this? I would like to integrate this to reap some startup benefits :)