@vnkozlov 2025-02-11
> I think this looks fine. The change is small, so we can update or remove it when migrating the SCCache into CDS.
> 
> @vnkozlov what do you think?

Yes, that is the plan.