@shipilev 2025-02-12
I think this PR would be superseded by @ashu-mehra's #39. We can still do it ahead of #39, removing the flag and accepting all platforms.