@shipilev 2025-02-12
/integrate